build: update test-only dependency - Jackson
Jackson is a transitive dependency of 'json-snapshot'. In order to
use Jackson as Java Module, we update it to the latest version instead
of patching it.

Unfortunately, 'json-snapshot' has not been updated in years and the
way we used it is not compatible with the latest Jackson version.

By adjusting the test setup as described here, we make it work:
https://github.com/json-snapshot/json-snapshot.github.io/issues/27#issuecomment-583307554

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>