build: replace Taskfile with Gradle tasks
The only things in the Taskfile that were still important are the
"update*" tasks. All other tasks just forwarded to Gradle.

With this change, the "update" tasks are implemented in Gradle so
Taskfile (and with that a separate tool) is no longer needed.

While testing the "updateProto" task, I discovered that we still
missed some adjustments in the "hedera -> hiero" change. These are
done in this PR as well.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>